+++ packages/twenty-front/src/modules/settings/admin-panel/components/SettingsAdminContent.tsx
@@ -1,9 +1,7 @@
-import { SettingsAdminEnvVariables } from '@/settings/admin-panel/components/SettingsAdminEnvVariables';
-import { SettingsAdminGeneral } from '@/settings/admin-panel/components/SettingsAdminGeneral';
+import { SettingsAdminTabContent } from '@/settings/admin-panel/components/SettingsAdminTabContent';
 import { SETTINGS_ADMIN_TABS } from '@/settings/admin-panel/constants/SettingsAdminTabs';
 import { SETTINGS_ADMIN_TABS_ID } from '@/settings/admin-panel/constants/SettingsAdminTabsId';
 import { TabList } from '@/ui/layout/tab/components/TabList';
-import { useTabList } from '@/ui/layout/tab/hooks/useTabList';
 import styled from '@emotion/styled';
 import { IconSettings2, IconVariable } from 'twenty-ui';
 
@@ -16,8 +14,6 @@ const StyledTabListContainer = styled.div`
 `;
 
 export const SettingsAdminContent = () => {
-  const { activeTabId } = useTabList(SETTINGS_ADMIN_TABS_ID);
-
   const tabs = [
     {
       id: SETTINGS_ADMIN_TABS.GENERAL,
@@ -31,27 +27,16 @@ export const SettingsAdminContent = () => {
     },
   ];
 
-  const renderContent = () => {
-    switch (activeTabId) {
-      case SETTINGS_ADMIN_TABS.GENERAL:
-        return <SettingsAdminGeneral />;
-      case SETTINGS_ADMIN_TABS.ENV_VARIABLES:
-        return <SettingsAdminEnvVariables />;
-      default:
-        return null;
-    }
-  };
-
   return (
     <>
       <StyledTabListContainer>
         <TabList
           tabs={tabs}
           tabListInstanceId={SETTINGS_ADMIN_TABS_ID}
-          behaveAsLinks={false}
+          behaveAsLinks={true}
         />
       </StyledTabListContainer>
-      {renderContent()}
+      <SettingsAdminTabContent />
     </>
   );
 };
