+++ packages/twenty-front/src/modules/settings/admin-panel/components/SettingsAdminEnvVariables.tsx
@@ -1,14 +1,7 @@
-import { SettingsAdminEnvVariablesRow } from '@/settings/admin-panel/components/SettingsAdminEnvVariablesRow';
-import { Table } from '@/ui/layout/table/components/Table';
-import { TableHeader } from '@/ui/layout/table/components/TableHeader';
-import { TableRow } from '@/ui/layout/table/components/TableRow';
+import { SettingsAdminEnvVariablesTable } from '@/settings/admin-panel/components/SettingsAdminEnvVariablesTable';
 import styled from '@emotion/styled';
 import { H1Title, H1TitleFontColor, Section } from 'twenty-ui';
 import { useGetEnvironmentVariablesGroupedQuery } from '~/generated/graphql';
-
-const StyledTable = styled(Table)`
-  margin-top: ${({ theme }) => theme.spacing(3)};
-`;
 
 const StyledGroupContainer = styled.div`
   margin-bottom: ${({ theme }) => theme.spacing(6)};
@@ -43,29 +36,10 @@ const StyledSubGroupDescription = styled.div``;
 const StyledGroupDescription = styled.div``;
 
 export const SettingsAdminEnvVariables = () => {
-  const { data: environmentVariables } =
-    useGetEnvironmentVariablesGroupedQuery();
-
-  const renderVariablesTable = (
-    variables: Array<{
-      name: string;
-      description: string;
-      value: string;
-      sensitive: boolean;
-    }>,
-  ) => (
-    <StyledTable>
-      <TableRow gridAutoColumns="4fr 3fr 2fr 1fr 1fr">
-        <TableHeader>Name</TableHeader>
-        <TableHeader>Description</TableHeader>
-        <TableHeader>Value</TableHeader>
-        <TableHeader align="right"></TableHeader>
-        <TableHeader align="right"></TableHeader>
-      </TableRow>
-      {variables.map((variable) => (
-        <SettingsAdminEnvVariablesRow key={variable.name} variable={variable} />
-      ))}
-    </StyledTable>
+  const { data: environmentVariables } = useGetEnvironmentVariablesGroupedQuery(
+    {
+      fetchPolicy: 'network-only',
+    },
   );
 
   return (
@@ -80,8 +54,9 @@ export const SettingsAdminEnvVariables = () => {
               </StyledGroupDescription>
             )}
             <StyledGroupVariablesContainer>
-              {group.variables.length > 0 &&
-                renderVariablesTable(group.variables)}
+              {group.variables.length > 0 && (
+                <SettingsAdminEnvVariablesTable variables={group.variables} />
+              )}
             </StyledGroupVariablesContainer>
             {group.subgroups.map((subgroup) => (
               <StyledSubGroupContainer key={subgroup.name}>
@@ -91,7 +66,9 @@ export const SettingsAdminEnvVariables = () => {
                     {subgroup.description}
                   </StyledSubGroupDescription>
                 )}
-                {renderVariablesTable(subgroup.variables)}
+                <SettingsAdminEnvVariablesTable
+                  variables={subgroup.variables}
+                />
               </StyledSubGroupContainer>
             ))}
           </StyledGroupContainer>
