+++ packages/twenty-front/src/modules/settings/admin-panel/states/adminPanelErrorState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const adminPanelErrorState = atom<string | null>({
+  key: 'adminPanelErrorState',
+  default: null,
+});
