+++ packages/twenty-front/src/modules/settings/admin-panel/states/userLookupResultState.ts
@@ -1,0 +1,7 @@
+import { UserLookup } from '@/settings/admin-panel/types/UserLookup';
+import { atom } from 'recoil';
+
+export const userLookupResultState = atom<UserLookup | null>({
+  key: 'userLookupResultState',
+  default: null,
+});
